+++ src/typeguard/__init__.py
@@ -46,6 +46,60 @@ _missing = object()
 
 T_CallableOrType = TypeVar('T_CallableOrType', bound=Callable[..., Any])
 
+# Lifted from mypy.sharedparse
+BINARY_MAGIC_METHODS = {
+    "__add__",
+    "__and__",
+    "__cmp__",
+    "__divmod__",
+    "__div__",
+    "__eq__",
+    "__floordiv__",
+    "__ge__",
+    "__gt__",
+    "__iadd__",
+    "__iand__",
+    "__idiv__",
+    "__ifloordiv__",
+    "__ilshift__",
+    "__imatmul__",
+    "__imod__",
+    "__imul__",
+    "__ior__",
+    "__ipow__",
+    "__irshift__",
+    "__isub__",
+    "__itruediv__",
+    "__ixor__",
+    "__le__",
+    "__lshift__",
+    "__lt__",
+    "__matmul__",
+    "__mod__",
+    "__mul__",
+    "__ne__",
+    "__or__",
+    "__pow__",
+    "__radd__",
+    "__rand__",
+    "__rdiv__",
+    "__rfloordiv__",
+    "__rlshift__",
+    "__rmatmul__",
+    "__rmod__",
+    "__rmul__",
+    "__ror__",
+    "__rpow__",
+    "__rrshift__",
+    "__rshift__",
+    "__rsub__",
+    "__rtruediv__",
+    "__rxor__",
+    "__sub__",
+    "__truediv__",
+    "__xor__",
+}
+
 
 class ForwardRefPolicy(Enum):
     """Defines how unresolved forward references are handled."""
@@ -65,12 +119,11 @@ class TypeHintWarning(UserWarning):
 
 
 class _TypeCheckMemo:
-    __slots__ = 'globals', 'locals', 'typevars'
+    __slots__ = 'globals', 'locals'
 
     def __init__(self, globals: Dict[str, Any], locals: Dict[str, Any]):
         self.globals = globals
         self.locals = locals
-        self.typevars = {}  # type: Dict[Any, type]
 
 
 def _strip_annotation(annotation):
@@ -169,11 +222,28 @@ def resolve_forwardref(maybe_ref, memo: _TypeCheckMemo):
 
 
 def get_type_name(type_):
-    name = getattr(type_, '__name__', None) or getattr(type_, '_name', None)
-    if name is None or name == 'Literal':
-        return str(type_)
-    else:
-        return name
+    name = (getattr(type_, '__name__', None) or getattr(type_, '_name', None) or
+            getattr(type_, '__forward_arg__', None))
+    if name is None:
+        origin = getattr(type_, '__origin__', None)
+        name = getattr(origin, '_name', None)
+        if name is None and not inspect.isclass(type_):
+            name = type_.__class__.__name__.strip('_')
+
+    args = getattr(type_, '__args__', ()) or getattr(type_, '__values__', ())
+    if args != getattr(type_, '__parameters__', ()):
+        if name == 'Literal':
+            formatted_args = ', '.join(str(arg) for arg in args)
+        else:
+            formatted_args = ', '.join(get_type_name(arg) for arg in args)
+
+        name = '{}[{}]'.format(name, formatted_args)
+
+    module = getattr(type_, '__module__', None)
+    if module not in (None, 'typing', 'typing_extensions', 'builtins'):
+        name = module + '.' + name
+
+    return name
 
 
 def find_function(frame) -> Optional[Callable]:
@@ -231,9 +301,9 @@ def function_name(func: Callable) -> str:
 
     """
     # For partial functions and objects with __call__ defined, __qualname__ does not exist
-    module = func.__module__
-    qualname = getattr(func, '__qualname__', repr(func))
-    return qualname if module == 'builtins' else '{}.{}'.format(module, qualname)
+    module = getattr(func, '__module__', '')
+    qualname = (module + '.') if module not in ('builtins', '') else ''
+    return qualname + getattr(func, '__qualname__', repr(func))
 
 
 def check_callable(argname: str, value, expected_type, memo: _TypeCheckMemo) -> None:
@@ -435,52 +505,57 @@ def check_class(argname: str, value, expected_type, memo: _TypeCheckMemo) -> Non
     if expected_type is Type:
         return
 
-    expected_class = None
-    if hasattr(expected_type, "__args__") and expected_type.__args__:
+    if getattr(expected_type, '__origin__', None) in (Type, type):
         expected_class = expected_type.__args__[0]
-    if expected_class:
-        if expected_class is Any:
-            return
-        elif isinstance(expected_class, TypeVar):
-            check_typevar(argname, value, expected_class, memo, True)
-        elif not issubclass(value, expected_class):
-            raise TypeError('{} must be a subclass of {}; got {} instead'.format(
-                argname, qualified_name(expected_class), qualified_name(value)))
+    else:
+        expected_class = expected_type
+
+    if expected_class is Any:
+        return
+    elif isinstance(expected_class, TypeVar):
+        check_typevar(argname, value, expected_class, memo, True)
+    elif getattr(expected_class, '__origin__', None) is Union:
+        for arg in expected_class.__args__:
+            try:
+                check_class(argname, value, arg, memo)
+                break
+            except TypeError:
+                pass
+        else:
+            formatted_args = ', '.join(get_type_name(arg) for arg in expected_class.__args__)
+            raise TypeError('{} must match one of the following: ({}); got {} instead'.format(
+                argname, formatted_args, qualified_name(value)
+            ))
+    elif not issubclass(value, expected_class):
+        raise TypeError('{} must be a subclass of {}; got {} instead'.format(
+            argname, qualified_name(expected_class), qualified_name(value)))
 
 
 def check_typevar(argname: str, value, typevar: TypeVar, memo: _TypeCheckMemo,
                   subclass_check: bool = False) -> None:
-    bound_type = resolve_forwardref(memo.typevars.get(typevar, typevar.__bound__), memo)
     value_type = value if subclass_check else type(value)
     subject = argname if subclass_check else 'type of ' + argname
-    if bound_type is None:
-        # The type variable hasn't been bound yet -- check that the given value matches the
-        # constraints of the type variable, if any
-        if typevar.__constraints__:
-            constraints = [resolve_forwardref(c, memo) for c in typevar.__constraints__]
-            if value_type not in constraints:
-                typelist = ', '.join(get_type_name(t) for t in constraints if t is not object)
-                raise TypeError('{} must be one of ({}); got {} instead'.
-                                format(subject, typelist, qualified_name(value_type)))
-    elif typevar.__covariant__ or typevar.__bound__:
+
+    if typevar.__bound__ is not None:
+        bound_type = resolve_forwardref(typevar.__bound__, memo)
         if not issubclass(value_type, bound_type):
             raise TypeError(
-                '{} must be {} or one of its subclasses; got {} instead'.
-                format(subject, qualified_name(bound_type), qualified_name(value_type)))
-    elif typevar.__contravariant__:
-        if not issubclass(bound_type, value_type):
-            raise TypeError(
-                '{} must be {} or one of its superclasses; got {} instead'.
-                format(subject, qualified_name(bound_type), qualified_name(value_type)))
-    else:  # invariant
-        if value_type is not bound_type:
-            raise TypeError(
-                '{} must be exactly {}; got {} instead'.
-                format(subject, qualified_name(bound_type), qualified_name(value_type)))
-
-    if typevar not in memo.typevars:
-        # Bind the type variable to a concrete type
-        memo.typevars[typevar] = value_type
+                '{} must be {} or one of its subclasses; got {} instead'
+                .format(subject, qualified_name(bound_type), qualified_name(value_type)))
+    elif typevar.__constraints__:
+        constraints = [resolve_forwardref(c, memo) for c in typevar.__constraints__]
+        for constraint in constraints:
+            try:
+                check_type(argname, value, constraint, memo)
+            except TypeError:
+                pass
+            else:
+                break
+        else:
+            formatted_constraints = ', '.join(get_type_name(constraint)
+                                              for constraint in constraints)
+            raise TypeError('{} must match one of the constraints ({}); got {} instead'
+                            .format(subject, formatted_constraints, qualified_name(value_type)))
 
 
 def check_literal(argname: str, value, expected_type, memo: _TypeCheckMemo):
@@ -589,6 +664,7 @@ def check_type(argname: str, value, expected_type, memo: Optional[_TypeCheckMemo
         (defaults to the calling frame's globals)
     :param locals: dictionary of local variables to use for resolving forward references
         (defaults to the calling frame's locals)
+    :raises TypeError: if there is a type mismatch
 
     """
     if expected_type is Any or isinstance(value, Mock):
@@ -680,6 +756,13 @@ def check_return_type(retval, memo: Optional[_CallMemo] = None) -> bool:
         try:
             check_type('the return value', retval, memo.type_hints['return'], memo)
         except TypeError as exc:  # suppress unnecessarily long tracebacks
+            # Allow NotImplemented if this is a binary magic method (__eq__() et al)
+            if retval is NotImplemented and memo.type_hints['return'] is bool:
+                # This does (and cannot) not check if it's actually a method
+                func_name = memo.func_name.rsplit('.', 1)[-1]
+                if len(memo.arguments) == 2 and func_name in BINARY_MAGIC_METHODS:
+                    return True
+
             raise TypeError(*exc.args) from None
 
     return True
@@ -773,7 +856,7 @@ class TypeCheckedAsyncGenerator:
         self.__send_type = rtype_args[1] if len(rtype_args) > 1 else Any
         self.__initialized = False
 
-    async def __aiter__(self):
+    def __aiter__(self):
         return self
 
     def __anext__(self):
@@ -845,12 +928,11 @@ def typechecked(func=None, *, always=False, _localns: Optional[Dict[str, Any]] =
             elif isinstance(attr, property):
                 kwargs = dict(doc=attr.__doc__)
                 for name in ("fset", "fget", "fdel"):
-                    property_func = getattr(attr, name)
-                    if property_func is None:
-                        continue
-                    kwargs[name] = typechecked(
-                        property_func, always=always, _localns=func.__dict__
-                    )
+                    property_func = kwargs[name] = getattr(attr, name)
+                    if property_func is not None and getattr(property_func, '__annotations__', ()):
+                        kwargs[name] = typechecked(
+                            property_func, always=always, _localns=func.__dict__
+                        )
 
                 setattr(func, key, property(**kwargs))